+++ CHANGELOG.md
@@ -18,6 +18,11 @@
 
 * Deprecate passing non-`deg` units to `color.hwb()`'s `$hue` argument.
 
+* Fix a number of bugs when determining whether selectors with pseudo-elements
+  are superselectors.
+
+* Treat `*` as a superselector of all selectors.
+
 ### Dart API
 
 * Add a top-level `fakeFromImport()` function for testing custom importers